requirements.txt
requests
http.client
python-dotenv
pytest